+++ lib/tuttle.rb
@@ -4,6 +4,9 @@ module Tuttle
 
   mattr_accessor :automount_engine
   mattr_accessor :enabled
+  mattr_accessor :track_notifications
+
+  @@automount_engine = @@enabled = @@track_notifications = false
 
   def self.setup
     yield self