+++ test/dummy/config/initializers/tuttle.rb
@@ -1,4 +1,5 @@
 Tuttle.setup do |config|
   config.enabled = true
   config.automount_engine = true
+  config.track_notifications = true
 end